Fix handleDM
