+++ src/hooks/notification.tsx
@@ -1,7 +1,9 @@
 import { useEffect } from 'react';
+import { useSetRecoilState } from 'recoil';
 import { useQueryClient } from '@tanstack/react-query';
 import { useSocketRef } from '@/hooks/chat';
 import { toast, ToastContentProps } from 'react-toastify';
+import { dmHistoryState } from '@/states/dmHistoryState';
 import NotificationGame from '@/molecule/NotificationGame';
 import NotificationChat from '@/molecule/NotificationChat';
 import NotificationDM from '@/molecule/NotificationDM';
@@ -12,6 +14,7 @@ import type { NotificationDMData } from '@/molecule/NotificationDM';
 export function useNotification() {
   const socketRef = useSocketRef(`ws://${import.meta.env.VITE_BASE_URL}:1234`);
   const queryClient = useQueryClient();
+  const setDmHistory = useSetRecoilState(dmHistoryState);
 
   useEffect(() => {
     const socket = socketRef.current;
@@ -41,6 +44,13 @@ export function useNotification() {
     };
     const handleDM = (data: NotificationDMData) => {
       // 최근 DM목록 업데이트
+      setDmHistory((currHistory) =>
+        currHistory.find(
+          (dmInfo) => dmInfo.opponent.user_id === data.sender.user_id,
+        )
+          ? currHistory
+          : [...currHistory, { opponent: data.sender, last_dm: new Date() }],
+      );
       toast(
         (props: ToastContentProps<NotificationDMData>) => (
           <NotificationDM {...props} />
@@ -60,5 +70,5 @@ export function useNotification() {
       socket.off('notification-chat', handleChat);
       socket.off('notification-dm', handleDM);
     };
-  }, [socketRef]);
+  }, [socketRef, setDmHistory]);
 }
